chore: move hextobin to common
