+++ common/tools/hextobin/build.sh
