+++ common/tools/hextobin/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "commander": "^3.0.0"
   },
-  "main": "build/hextobin.js",
+  "main": "build/index.js",
   "bin": {
     "hextobin": "build/hextobin.js"
   },
